Revert "Test/cli support"
