--- .env.sample
@@ -1,5 +1,0 @@
-# Test Connection
-HOST=
-PORT=22
-USER=
-PASSWORD=
