+++ .github/workflows/webpack.yaml
@@ -11,8 +11,6 @@ on:
 
 jobs:
   build:
-    name: Webpack
-  
     runs-on: ubuntu-latest
 
     strategy:
