--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "src/test/testFixtures/ibmi-company_system"]
-	path = src/test/testFixtures/ibmi-company_system
-	url = https://github.com/IBM/ibmi-company_system.git
