+++ src/testing/suites/decorationProvider.ts
