+++ src/testing/suites/jobLogCommand.ts
