+++ types/package.json
@@ -17,7 +17,7 @@
   },
   "typings": "./ibmiProjectExplorer.d.ts",
   "scripts": {
-    "prepublish": "rm -rf test",
+    "prepublish": "rm -rf testing",
     "deploy": "npm publish --access public"
   },
   "dependencies": {
